# YJIT: Use registers to pass stack temps to C calls