## Background

The current `asm.ccall` expects `asm.spill_temps()` to be called outside it. It was necessary when some arguments were popped before a C call and using registers for them would cause stack temp registers and argument registers to depend on each other.

## Changes

https://github.com/ruby/ruby/pull/7918 can automatically remove such issues from C arguments, so this PR changes `asm.ccall` to lazily spill stack temps that are still on stack and avoid spilling stack temps that are already popped. This not only makes the code cleaner but also optimizes the following methods by removing memory reads and writes:

* `Array#[]`
* `Module#===`
* `Integer#/`
* `Integer#[]`
* `Integer#%`
* `String#bytesize`
* `Struct#[]=`

Many other methods use `jit_prepare_routine_call` and it's generally not safe to skip spilling stack temps in such cases because arguments also need to be on the VM stack for GC. We could still let them use registers in arguments even after spilling them in some cases, but it's out of scope in this PR.

## Benchmark
### Headline
liquid-compile and mail reliably speed up by 2%.

```
before: ruby 3.3.0dev (2023-06-07T01:10:08Z master 46583f7dcf) +YJIT [x86_64-linux]
after: ruby 3.3.0dev (2023-06-08T23:46:17Z yjit-temp-ccall ccdd408693) +YJIT [x86_64-linux]

--------------  -----------  ----------  ----------  ----------  -------------  ------------
bench           before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
activerecord    32.3         5.4         32.0        5.8         1.01           1.01
chunky_png      420.2        0.7         417.8       0.5         1.00           1.01
erubi_rails     11.5         26.3        11.5        28.4        0.95           1.00
hexapdf         1358.5       2.0         1365.5      0.9         0.99           0.99
liquid-c        41.2         1.2         40.6        1.4         0.98           1.01
liquid-compile  37.9         2.6         37.1        3.0         0.95           1.02
liquid-render   73.4         0.6         73.2        0.7         0.99           1.00
mail            90.5         0.3         88.4        0.2         0.98           1.02
psych-load      1232.1       0.3         1221.6      0.2         1.01           1.01
railsbench      1243.6       1.4         1242.8      1.4         0.99           1.00
ruby-lsp        38.6         19.7        38.3        21.0        0.98           1.01
sequel          52.0         1.9         51.8        2.3         1.00           1.00
--------------  -----------  ----------  ----------  ----------  -------------  ------------
```

### Other
A number of them show a speedup. fannkuchredux is 7% faster, and nbody is 5% faster.

```
before: ruby 3.3.0dev (2023-06-07T01:10:08Z master 46583f7dcf) +YJIT [x86_64-linux]
after: ruby 3.3.0dev (2023-06-08T23:46:17Z yjit-temp-ccall ccdd408693) +YJIT [x86_64-linux]

-------------  -----------  ----------  ----------  ----------  -------------  ------------
bench          before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
binarytrees    137.0        1.4         137.3       1.4         1.02           1.00
erubi          161.0        0.3         156.8       0.1         1.02           1.03
etanni         261.3        0.6         255.4       0.3         1.02           1.02
fannkuchredux  472.0        0.3         439.9       0.3         1.01           1.07
lee            611.3        1.1         607.7       1.2         1.01           1.01
nbody          45.7         0.6         43.5        0.3         1.04           1.05
optcarrot      1400.7       0.8         1390.9      0.9         1.02           1.01
ruby-json      2428.6       0.1         2499.8      0.2         0.97           0.97
rubykon        4409.4       0.3         4265.8      0.3         1.03           1.03
-------------  -----------  ----------  ----------  ----------  -------------  ------------
```